Cars table and Clients Cars List with Hibernate - Done

+++ src/main/java/com/example/garage/service/ClientService.java
@@ -23,11 +23,13 @@ public class ClientService {
         return clientRepository.findAll();
     }
 
-    public Optional<Client> getClientById(int id) {
+    public Optional<Client> getClientById(long id) {
+//        System.out.println("!!!!!!!!!!!!!!!!!!!!      " + clientRepository.findById((long)id));
         return clientRepository.findById(Long.valueOf(id));
     }
 
     public Client saveClient(Client client) {
+//        System.out.println(client.getCars());
         return clientRepository.save(client);
     }
